refactor: remove request for data in header
