+++ src/components/header/header.tsx
@@ -1,7 +1,4 @@
-import { useEffect } from "react";
-
-import { fetchTopCoinsData } from "../../redux/actions/get-coins-asynk-thunk";
-import { useAppDispatch, useAppSelector } from "../../hooks";
+import { useAppSelector } from "../../hooks";
 
 import { Cart, PopularCoin } from "./components";
 
@@ -9,12 +6,6 @@ import styles from "./header.module.scss";
 
 export const Header: React.FC = () => {
   const { headerCoin } = useAppSelector((state) => state.coins);
-
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    dispatch(fetchTopCoinsData(3));
-  }, [dispatch]);
 
   return (
     <>
