+++ src/main.tsx
@@ -1,17 +1,17 @@
-import { StrictMode } from 'react';
-import { createRoot } from 'react-dom/client';
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
 
-import { RouterProvider } from 'react-router-dom';
-import { router } from './routes/routes.tsx';
+import { RouterProvider } from "react-router-dom";
+import { router } from "./routes/routes.tsx";
 
-import './styles/index.scss';
-import { Provider } from 'react-redux';
-import store from './redux/store.ts';
+import "./styles/index.scss";
+import { Provider } from "react-redux";
+import store from "./redux/store.ts";
 
-createRoot(document.getElementById('root')!).render(
+createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <Provider store={store}>
       <RouterProvider router={router} />
     </Provider>
-  </StrictMode>,
+  </StrictMode>
 );
